# leargon-backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Léargon API
  description: 'REST API for Léargon data and process governance platform.


    ## Authentication

    The API uses JWT (JSON Web Token) based authentication. After logging in or signing up, you will receive an access token
    that must be included in the `Authorization` header as a Bearer token for protected endpoints.


    ## Authorization Roles

    - **ROLE_USER**: Standard user role, assigned to all users by default

    - **ROLE_ADMIN**: Administrative role with full user management capabilities


    ## Fallback Administration Protection

    The system includes a fallback admin user that cannot be modified or deleted through the API. Any attempt to modify this
    user will result in a 403 Forbidden response.

    '
  version: 1.0.0
  contact:
    name: Léargon
servers:
- url: http://localhost:8081
  description: Development server
- url: http://localhost:8080
  description: Local container server
...
paths:
...
  /business-entities:
    get:
      tags:
      - BusinessEntity
      summary: Get all entities
      description: Returns a list of all entities with their translations.
      operationId: getAllBusinessEntities
      security:
      - bearerAuth: []
      responses:
        '200':
          description: List of entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessEntityResponse'
        '401':
          description: Not authenticated
    post:
      tags:
      - BusinessEntity
      summary: Create an entity
      description: Creates a new entity with translations. The creator becomes the Data Owner by default.
      operationId: createBusinessEntity
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusinessEntityRequest'
      responses:
        '201':
          description: BusinessEntity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/tree:
    get:
      tags:
      - BusinessEntity
      summary: Get business entity hierarchy
      description: Returns business entities as a hierarchical tree structure.
      operationId: getBusinessEntityTree
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Hierarchical business entity tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessEntityTreeResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}:
    get:
      tags:
      - BusinessEntity
      summary: Get business entity by key
      description: Returns detailed information about a specific business entity including all translations.
      operationId: getBusinessEntityByKey
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: BusinessEntity key (e.g., 'sales.retail.customer')
        schema:
          type: string
      responses:
        '200':
          description: BusinessEntity details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '404':
          description: BusinessEntity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
    delete:
      tags:
      - BusinessEntity
      summary: Delete business entity
      description: Permanently deletes an business entity. Only the Data Owner or an Administrator can delete an entity.
      operationId: deleteBusinessEntity
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      responses:
        '204':
          description: Business Entity deleted successfully
        '403':
          description: Only the Data Owner or an Administrator can delete this entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/classifications:
    put:
      tags:
      - BusinessEntity
      summary: Assign classifications to business entity
      description: Replaces all classification assignments on a business entity. Each classification must be assignable to
        BUSINESS_ENTITY and only one value per classification is allowed.
      operationId: assignClassificationsToEntity
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ClassificationAssignmentRequest'
      responses:
        '200':
          description: Classifications assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '400':
          description: Invalid assignment (wrong assignable type, invalid value, duplicate classification)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only the Data Owner or an Administrator can assign classifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity, classification, or value not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/data-owner:
    put:
      tags:
      - BusinessEntity
      summary: Update business entity data owner
      description: Updates the data owner of a business entity. Only the Data Owner or an Administrator can edit an entity.
      operationId: updateBusinessEntityDataOwner
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBusinessEntityDataOwnerRequest'
      responses:
        '200':
          description: Business Entity data owner updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '403':
          description: Only the Data Owner or an Administrator can edit this entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/descriptions:
    put:
      tags:
      - BusinessEntity
      summary: Update business entity descriptions
      description: Replaces all descriptions of a business entity. Only the Data Owner or an Administrator can edit an entity.
      operationId: updateBusinessEntityDescriptions
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/LocalizedText'
      responses:
        '200':
          description: Business Entity descriptions updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only the Data Owner or an Administrator can edit this entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/domain:
    put:
      tags:
      - BusinessEntity
      summary: Assign business domain to business entity
      description: Assigns a business domain to an business entity. Only the Data Owner or an Administrator can assign domains.
      operationId: assignBusinessDomainToBusinessEntity
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignBusinessDomainRequest'
      responses:
        '200':
          description: Business Domain assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only the Data Owner or an Administrator can assign domains
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity or domain not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/interfaces:
    put:
      tags:
      - BusinessEntity
      summary: Update business entity interfaces
      description: Updates the interface entities of a business entity. Only the Data Owner or an Administrator can edit an
        entity.
      operationId: updateBusinessEntityInterfaces
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBusinessEntityInterfacesRequest'
      responses:
        '200':
          description: Business Entity interfaces updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only the Data Owner or an Administrator can edit this entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity or interface entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/localized:
    get:
      tags:
      - BusinessEntity
      summary: Get localized business entity
      description: Returns a single-locale view of a business entity. If no locale parameter given, uses the user's preferred
        language, falling back to the DB default locale.
      operationId: getLocalizedBusinessEntity
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      - name: locale
        in: query
        required: false
        description: Locale code (e.g., 'en', 'de'). Defaults to user's preferred language.
        schema:
          type: string
      responses:
        '200':
          description: Localized business entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalizedBusinessEntityResponse'
        '404':
          description: Business Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/names:
    put:
      tags:
      - BusinessEntity
      summary: Update business entity names
      description: Replaces names of a business entity. Only the Data Owner or an Administrator can edit an entity.
      operationId: updateBusinessEntityNames
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/LocalizedText'
      responses:
        '200':
          description: Business Entity names updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only the Data Owner or an Administrator can edit this entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/parent:
    put:
      tags:
      - BusinessEntity
      summary: Update business entity parent
      description: Updates the parent of a business entity. Only the Data Owner or an Administrator can edit an entity.
      operationId: updateBusinessEntityParent
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBusinessEntityParentRequest'
      responses:
        '200':
          description: Business Entity parent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '400':
          description: Invalid request data (e.g., would create cycle)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only the Data Owner or an Administrator can edit this entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity or parent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/relationships:
    post:
      tags:
      - BusinessEntity
      summary: Create a relationship
      description: Creates a relationship between this entity and another entity. Only the Data Owner or an Administrator
        can create relationships.
      operationId: createBusinessEntityRelationship
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key (first entity)
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusinessEntityRelationshipRequest'
      responses:
        '201':
          description: Relationship created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only the Data Owner or an Administrator can create relationships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/relationships/{relationshipId}:
    put:
      tags:
      - BusinessEntity
      summary: Update a relationship
      description: Updates cardinalities and descriptions of a relationship. Only the Data Owner or an Administrator can update
        relationships.
      operationId: updateBusinessEntityRelationship
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      - name: relationshipId
        in: path
        required: true
        description: Relationship ID
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBusinessEntityRelationshipRequest'
      responses:
        '200':
          description: Relationship updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only the Data Owner or an Administrator can update relationships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity or relationship not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
    delete:
      tags:
      - BusinessEntity
      summary: Delete a relationship
      description: Deletes a relationship between entities. Only the Data Owner or an Administrator can delete relationships.
      operationId: deleteBusinessEntityRelationship
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      - name: relationshipId
        in: path
        required: true
        description: Relationship ID
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: Relationship deleted successfully
        '403':
          description: Only the Data Owner or an Administrator can delete relationships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business Entity or relationship not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/versions:
    get:
      tags:
      - BusinessEntity
      summary: Get business entity version history
      description: Returns the version history of an business entity.
      operationId: getVersions
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      responses:
        '200':
          description: Business Entity Version history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessEntityVersionResponse'
        '404':
          description: Business Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/versions/{versionNumber}/diff:
    get:
      tags:
      - BusinessEntity
      summary: Get diff between versions
      description: Returns the diff between a specific version and the previous version.
      operationId: getVersionDiff
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      - name: versionNumber
        in: path
        required: true
        description: Version number to compare (compares with previous version)
        schema:
          type: integer
      responses:
        '200':
          description: Version diff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionDiffResponse'
        '404':
          description: BusinessEntity or version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
  /business-entities/{key}/retention-period:
    put:
      tags:
      - BusinessEntity
      summary: Update retention period
      description: Sets or clears the retention period for a business entity. Requires data owner or admin.
      operationId: updateBusinessEntityRetentionPeriod
      security:
      - bearerAuth: []
      parameters:
      - name: key
        in: path
        required: true
        description: Business Entity key
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRetentionPeriodRequest'
      responses:
        '200':
          description: Updated business entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessEntityResponse'
        '403':
          description: Forbidden (not data owner or admin)
        '404':
          description: Business entity not found
        '401':
          description: Not authenticated
...
  /users/me:
    get:
      tags:
      - User
      summary: Get current user profile
      description: Returns the profile information of the currently authenticated user.
      operationId: getCurrentUser
      security:
      - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /authentication/login or /authentication/signup
  schemas:
...
    ClassificationAssignmentRequest:
      type: object
      required:
      - classificationKey
      - valueKey
      properties:
        classificationKey:
          type: string
          description: Key of the classification
        valueKey:
          type: string
          description: Key of the value within the classification
...
    CreateBusinessEntityRequest:
      type: object
      required:
      - names
      properties:
        names:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LocalizedText'
          description: Names for the entity (at least one required)
        descriptions:
          type: array
          items:
            $ref: '#/components/schemas/LocalizedText'
          description: Descriptions for the entity
        dataOwnerUsername:
          type: string
          nullable: true
          description: Username of the data owner (defaults to creator if not specified)
        parentKey:
          type: string
          nullable: true
          description: Parent entity key (null for top-level entity)
        interfaces:
          type: array
          items:
            type: string
            nullable: false
            description: Interface entity key
        retentionPeriod:
          type: string
          nullable: true
          description: Retention period description for the entity
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: User not found
        path:
          type: string
          description: Request path
          example: /administration/users/999
        status:
          type: integer
          description: HTTP status code
          example: 404
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2024-01-15T10:30:00Z'
...